skills/k8s-gke/requirements.txt
kubernetes
google-cloud-container
